move the jitter up function
